vmebus: compile for 64-bit architectures
This patch contains the necessary changes to compile the vmebridge driver
for x86_64.

* get rid of old makefiles for good. We'll do this slowly until
  we have a normal build for *everything*. At the very least, this
  compiles for L865 (both operational kernels) and L866. It has been
  reported even to work with a CTRV driver on top.

* fix the compat_semaphore hack into a normal semaphore. The PI
  semaphore will work (and is needed) with current operational
  -rt2[17] kernels. It's not needed after 2.6.31.

* fix a parameter that should be a pointer to unsigned long

+++ vmebridge/driver/vme_window.c
@@ -413,7 +413,7 @@ struct vme_mapping* find_vme_mapping_from_addr(unsigned logaddr)
 		if (mutex_lock_interruptible(&window->lock))
 			return NULL;
 		list_for_each_entry(mapping, &window->mappings, list) {
-			if ((unsigned)mapping->desc.kernel_va == logaddr) {
+			if ((unsigned long)mapping->desc.kernel_va == logaddr) {
 				mutex_unlock(&window->lock);
 				return &mapping->desc; /* bingo */
 			}
